decouples router and inspector metrics
